% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_calendar_dim.R
\name{build_calendar_dim}
\alias{build_calendar_dim}
\title{Builds a \code{calendar_dim} table on \code{oliver_replica}}
\format{The columns are defined in \code{\link{meta_calendar_dim}}.}
\source{
Enrico Service - a free service written in PHP providing public holidays for several countries.
}
\usage{
build_calendar_dim()
}
\arguments{
\item{from_date}{The starting date for the table to be built.}

\item{to_date}{The stopping date for the table to be built.}

\item{country_filter}{The country to be used for the holidays.}

\item{region_filter}{The region or state to be used for the holidays.}

\item{tz}{The time zone to be used for all of the dates.}

\item{wrt_sch_name}{Optional. The name of the schema used to write the table.
Only required when \code{establish_con=TRUE}.}

\item{establish_con}{Required. A boolean flag indicating whether or not to run.}
}
\description{
A calendar table can be immensely useful, particularly for reporting purposes
and for determining things like business days between two dates. This function
builds a relatively static calendar table on \code{oliver_replica}.

Holidays are determined on the basis of the Enrico holiday service, which is
available at the following link: \url{http://www.kayaposoft.com/enrico/}.
The resulting table also contains a flag (\code{fl_pre_timeline}) which can be used to determine
whether or not a date takes place before or after the parent-child-visitation (PCV)
timeline feature which was released on 2017-10-20.
}
\examples{
\dontrun{

build_calendar_dim()
}
}
\author{
mienko
}
