% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_visitation_referral_attribute_dim.R
\name{build_visitation_referral_attribute_dim}
\alias{build_visitation_referral_attribute_dim}
\title{Builds a \code{visitation_referral_attribute_dim} table on \code{oliver_replica}}
\format{The columns are defined in \code{\link{meta_visitation_referral_attribute_dim}}.}
\source{
\code{oliver_replica}
}
\usage{
build_visitation_referral_attribute_dim()
}
\arguments{
\item{bld_sch_name}{Optional. The name of the schema used to build the table.
Only required when \code{establish_con=TRUE}.}

\item{wrt_sch_name}{Optional. The name of the schema used to write the table.
Only required when \code{establish_con=TRUE}.}

\item{establish_con}{Required. A boolean flag indicating whether or not to run
\code{\link{establish_con_olvr_rplc}} before building the table. (Defaults to \code{FALSE})}
}
\description{
This table is designed to provide contextual (i.e. dimensional) information to the
\code{visitation_referral_fact} table on the \code{oliver_replica} database.
}
\examples{
\dontrun{

build_visitation_referral_attribute_dim()
}
}
\author{
mienko
}
