% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_visitation_referral_fact.R
\name{build_visitation_referral_fact}
\alias{build_visitation_referral_fact}
\title{Builds a \code{visitation_referral_fact} table on \code{oliver_replica}}
\format{The columns are defined in \code{\link{meta_visitation_referral_fact}}.}
\source{
\code{oliver_replica}
}
\usage{
build_visitation_referral_fact()
}
\arguments{
\item{bld_sch_name}{Optional. The name of the schema used to build the table.
Only required when \code{establish_con=TRUE}.}

\item{wrt_sch_name}{Optional. The name of the schema used to write the table.
Only required when \code{establish_con=TRUE}.}

\item{establish_con}{Required. A boolean flag indicating whether or not to run
\code{\link{establish_con_olvr_rplc}} before building the table. (Defaults to \code{FALSE})}
}
\description{
This table is designed to provide basic information related to a supported
visitation referral and key dates associated with this referral. The table also links to
dimensional tables which provide additional contextual information about the referral.
}
\examples{
\dontrun{

build_visitation_referral_fact()
}
}
\author{
mienko
}
