% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_comment_olvr_rplc.R
\name{col_comment_olvr_rplc}
\alias{col_comment_olvr_rplc}
\title{Comment on a table column through a connection to \code{oliver_replica}}
\usage{
col_comment_olvr_rplc(sch_name = NA, establish_con = FALSE, tbl_name = NA,
  col_name = NA, col_comment = NA)
}
\arguments{
\item{sch_name}{Optional. The name of the schema with the column to be commented.
Only required when \code{establish_con=TRUE}.}

\item{establish_con}{Required. A boolean flag indicating whether or not to run
\code{\link{establish_con_olvr_rplc}} before adding comments. (Defaults to \code{FALSE})}

\item{tbl_name}{Required. The name of the table with the column to be commented.}

\item{col_name}{Required. The name of the column to be commented.}

\item{col_comment}{Optional. If left unspecified, the column comment will be set to \code{null}.}
}
\description{
Adds comments to table columns within the \code{oliver_replica}. These comments
are then used to document various functions within the oliveR2 package.
}
\examples{
\dontrun{
col_comment_olvr_rplc(sch_name = "independent"
 ,tbl_name = "calendar_dim"
 ,col_name = "id_calendar_dim"
 ,col_comment = "My Comment")
 }
}
