% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_diff_olvr_rplc.R
\name{date_diff_olvr_rplc}
\alias{date_diff_olvr_rplc}
\title{Builds a tibble based on state transitions in \code{visitation_referral_fact}}
\usage{
date_diff_olvr_rplc(start = "assigned", stop = "scheduled",
  contract_network = 2, obs_window_start = ymd(20170301),
  obs_window_stop = NA, bld_sch_name = "independent",
  establish_con = TRUE, exclude_negative = TRUE, target = 3,
  weekend_holiday_credit = TRUE, jitter = FALSE)
}
\arguments{
\item{start}{Required. A character string to be passed to \code{dplyr::contains()}. All column names in the
\code{visitation_referral_fact} table will be searched for the value to indicate the "start" column(s).}

\item{stop}{Required. A character string to be passed to \code{dplyr::contains()}. All column names in the
\code{visitation_referral_fact} table will be searched for the value to indicate the "stop" column(s).}

\item{contract_network}{An integer value indicating the Oliver provider id for the network administrator (if any).
This value may be set to \code{NA} to include all providers, regardles of their contract status. The default is 2, the Family Impact Network (FIN).}

\item{obs_window_start}{A \code{lubridate} date value indicating the lower bound for the column of starts. Defaults to \code{ymd(20170301)}. The this parameter can be set to \code{NA} to indicate no bound.}

\item{obs_window_stop}{A \code{lubridate} date value indicating the upper bound for the column of starts. Defaults to \code{now()}. The this parameter can be set to \code{NA} to indicate no bound.}

\item{bld_sch_name}{Optional. The name of the schema used to build the tibble.
Only required when \code{establish_con=TRUE}.}

\item{establish_con}{Required. A boolean flag indicating whether or not to run.}

\item{exclude_negative}{Required. Should negative periods be set to NA? Defaults to TRUE.}

\item{target}{Optional. A \code{lubridate} period object which specifies the period of time at target for a
particular period measurement.}

\item{jitter}{Required. Should the results be jittered using \code{jitter()}? Defaults to FALSE.}
}
\value{
A tibble object.
}
\description{
Builds a tibble based on state transitions in \code{visitation_referral_fact}
}
