% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_con_olvr_rplc.R
\name{establish_con_olvr_rplc}
\alias{establish_con_olvr_rplc}
\title{Establish a connection to \code{oliver_replica}}
\usage{
establish_con_olvr_rplc(set_schema = "staging", connection_name = "con")
}
\arguments{
\item{set_schema}{The starting date for the table to be built.}

\item{connection_name}{The stopping date for the table to be built.}
}
\description{
Searches for \code{oliver_replica} environment variables and passes these to
\code{\link[DBI]{dbConnect}}. Allows the user to specify the search path (i.e. schema)
to use in the connection and name the connection (which will be written to \code{.GlobalEnv}).
}
\note{
The function will first search \code{.GlobalEnv} for an object matching
\code{connection_name}. If present, the connection will be dropped and overwritten with
a new connection to \code{oliver_replica}.
}
\examples{

establish_con_olvr_rplc(set_schema = 'MySchema', connection_name = "MyCon")

}
