% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_calendar_dim}
\alias{meta_calendar_dim}
\title{Metadata for calendar_dim}
\format{The table contains two columns:
\describe{
  \item{col_desc}{the column description in the database table}
  \item{carat}{the column name in the database table}
}}
\source{
\code{\link{build_calendar_dim}}
}
\usage{
data(meta_calendar_dim)
}
\description{
A data frame decribing the \code{calendar_dim} table on \code{oliver_replica}.
}
\keyword{datasets}
