% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_col_comment_olvr_rplc.R
\name{retrieve_col_comment_olvr_rplc}
\alias{retrieve_col_comment_olvr_rplc}
\title{Retreive comments on a table column through a connection to \code{oliver_replica}}
\usage{
retrieve_col_comment_olvr_rplc(sch_name = NA, establish_con = FALSE,
  tbl_name = NA, col_idx = NA)
}
\arguments{
\item{sch_name}{Optional. The name of the schema with the column with comments to be retreived.
Only required when \code{establish_con=TRUE}.}

\item{establish_con}{Required. A boolean flag indicating whether or not to run
\code{\link{establish_con_olvr_rplc}} before adding comments. (Defaults to \code{FALSE})}

\item{tbl_name}{Required. The name of the table with the column to be commented.}

\item{col_idx}{Required. The index of the column with comments to be retreived (\code{1:length(tbl)}).}
}
\description{
Retreives comments to table columns within the \code{oliver_replica}. These comments
may have been placed by \code{\link{col_comment_olvr_rplc}}, but the function will
retreive any comments existing in the table, regardless of how they were created.
}
\examples{
\dontrun{
retrieve_col_comment_olvr_rplc(sch_name = "independent"
 ,establish_con = TRUE
 ,tbl_name = "calendar_dim"
 ,col_idx = 1)
 }
}
